Initial setup for localization.
Created locale and localizable folders with localizations for
English and Portuguese.
Added two terms to be localizable just for testing.
Changed name of _about partial to _about_partial because it seemed
to be conflicting with other file named about.html.erb.
Moved all html.erb files to localizable folder. Partials stayes in
the root folder.